fixed dot comma issue and reset _running variable when run stopped

+++ AnalogOutput/AnalogOutput/Data/DataChannel.cs
@@ -10,6 +10,6 @@ namespace AnalogOutput.Data
     {               
         public DataStep[] Steps;
         public string Name = "";
-        public float InitialValue;
+        public double InitialValue;
     }
 }
